Prepare for Ares-local handling in src/ares. Move AdjacencyBonus to src/ares.
