+++ src/ISpace.ts
@@ -3,7 +3,7 @@ import { SpaceBonus } from "./SpaceBonus";
 import { SpaceType } from "./SpaceType";
 import { ITile } from "./ITile";
 import { Player } from "./Player";
-import { AdjacencyBonus } from "./cards/ares/AdjacencyBonus";
+import { AdjacencyBonus } from "./ares/AdjacencyBonus";
 
 export interface ISpace {
     id: string;
