+++ src/cards/CommercialDistrict.ts
@@ -9,7 +9,7 @@ import {ISpace} from '../ISpace';
 import { Resources } from '../Resources';
 import { CardName } from '../CardName';
 import { Board } from '../Board';
-import { AdjacencyBonus } from "./ares/AdjacencyBonus";
+import { AdjacencyBonus } from "../ares/AdjacencyBonus";
 
 export class CommercialDistrict implements IProjectCard {
     public cost: number = 16;
