+++ src/cards/ares/IndustrialCenterAres.ts
@@ -1,7 +1,7 @@
 import { CardName } from "../../CardName";
 import { SpaceBonus } from "../../SpaceBonus";
 import { IndustrialCenter } from "../IndustrialCenter";
-import { AdjacencyBonus } from "./AdjacencyBonus";
+import { AdjacencyBonus } from "../../ares/AdjacencyBonus";
 
 export class IndustrialCenterAres extends IndustrialCenter {
   public name: CardName = CardName.INDUSTRIAL_CENTER_ARES;
