+++ src/cards/ares/LavaFlowsAres.ts
@@ -1,7 +1,7 @@
 import { CardName } from "../../CardName";
 import { SpaceBonus } from "../../SpaceBonus";
 import { LavaFlows } from "../LavaFlows";
-import { AdjacencyBonus } from "./AdjacencyBonus";
+import { AdjacencyBonus } from "../../ares/AdjacencyBonus";
 
 export class LavaFlowsAres extends LavaFlows {
   public name: CardName = CardName.LAVA_FLOWS_ARES;
