+++ src/cards/ares/MoholeAreaAres.ts
@@ -1,7 +1,7 @@
 import { CardName } from "../../CardName";
 import { SpaceBonus } from "../../SpaceBonus";
 import { MoholeArea } from "../MoholeArea";
-import { AdjacencyBonus } from "./AdjacencyBonus";
+import { AdjacencyBonus } from "../../ares/AdjacencyBonus";
 
 export class MoholeAreaAres extends MoholeArea {
   public name: CardName = CardName.MOHOLE_AREA_ARES;
