+++ src/cards/ares/NaturalPreserveAres.ts
@@ -1,6 +1,6 @@
 import { CardName } from "../../CardName";
 import { NaturalPreserve } from "../NaturalPreserve";
-import { AdjacencyBonus, AresSpaceBonus } from "./AdjacencyBonus";
+import { AdjacencyBonus, AresSpaceBonus } from "../../ares/AdjacencyBonus";
 
 export class NaturalPreserveAres extends NaturalPreserve {
   public name: CardName = CardName.NATURAL_PRESERVE_ARES;
