+++ tsconfig.json
@@ -23,6 +23,7 @@
   },
   "include": [
     "src/*.ts",
+    "src/ares/*.ts",
     "src/components/*.ts",
     "src/directives/*.ts",
     "src/cards/*.ts",
@@ -41,7 +42,7 @@
     "src/interrupts/*.ts",
     "src/database/*.ts",
     "src/turmoil/*.ts",
-    "src/turmoil/globalEvents*.ts",    
+    "src/turmoil/globalEvents*.ts",  
   ],
   "files": [
     "script.ts",
